+++ src/Blitz.php
@@ -10,6 +10,7 @@ use craft\base\Element;
 use craft\base\Plugin;
 use craft\elements\User;
 use craft\events\BatchElementActionEvent;
+use craft\events\DefineHtmlEvent;
 use craft\events\DeleteElementEvent;
 use craft\events\ElementEvent;
 use craft\events\MoveElementEvent;
@@ -40,6 +41,7 @@ use putyourlightson\blitz\drivers\deployers\BaseDeployer;
 use putyourlightson\blitz\drivers\generators\BaseCacheGenerator;
 use putyourlightson\blitz\drivers\purgers\BaseCachePurger;
 use putyourlightson\blitz\drivers\storage\BaseCacheStorage;
+use putyourlightson\blitz\helpers\ElementSidebarHelper;
 use putyourlightson\blitz\helpers\IntegrationHelper;
 use putyourlightson\blitz\helpers\RefreshCacheHelper;
 use putyourlightson\blitz\models\RefreshDataModel;
@@ -111,7 +113,7 @@ class Blitz extends Plugin
     /**
      * @inheritdoc
      */
-    public string $schemaVersion = '4.13.0';
+    public string $schemaVersion = '4.22.0';
 
     /**
      * @inheritdoc
@@ -153,6 +155,7 @@ class Blitz extends Plugin
             $this->registerCpUrlRules();
             $this->registerUtilities();
             $this->registerWidgets();
+            $this->registerSidebarPanels();
             $this->registerRedirectAfterInstall();
 
             if (Craft::$app->getEdition() === Craft::Pro) {
@@ -550,6 +553,22 @@ class Blitz extends Plugin
     }
 
     /**
+     * Registers sidebar panels
+     *
+     * @since 4.22.0
+     */
+    private function registerSidebarPanels(): void
+    {
+        Event::on(Element::class, Element::EVENT_DEFINE_SIDEBAR_HTML,
+            function(DefineHtmlEvent $event) {
+                /** @var Element $element */
+                $element = $event->sender;
+                $event->html .= ElementSidebarHelper::getSidebarHtml($element);
+            },
+        );
+    }
+
+    /**
      * Registers redirect after install
      */
     private function registerRedirectAfterInstall(): void
@@ -607,6 +626,12 @@ class Blitz extends Plugin
                         ],
                         'blitz:refresh-tagged' => [
                             'label' => Craft::t('blitz', 'Refresh tagged cache'),
+                        ],
+                        'blitz:refresh-page' => [
+                            'label' => Craft::t('blitz', 'Refresh pages via the sidebar panel and diagnostics utility'),
+                        ],
+                        'blitz:view-sidebar-panel' => [
+                            'label' => Craft::t('blitz', 'View sidebar panel on element edit pages'),
                         ],
                     ],
                 ];