+++ src/helpers/DiagnosticsHelper.php
@@ -63,18 +63,12 @@ class DiagnosticsHelper
 
     public static function getPagesCount(int $siteId): int
     {
-        return CacheRecord::find()
-            ->where(['siteId' => $siteId])
-            ->andWhere(['not', self::IS_CACHED_INCLUDE_CONDITION])
-            ->count();
+        return self::getPagesQuery($siteId)->count();
     }
 
     public static function getIncludesCount(int $siteId): int
     {
-        return CacheRecord::find()
-            ->where(['siteId' => $siteId])
-            ->andWhere(self::IS_CACHED_INCLUDE_CONDITION)
-            ->count();
+        return self::getIncludesQuery($siteId)->count();
     }
 
     public static function getParamsCount(int $siteId): int
@@ -218,7 +212,7 @@ class DiagnosticsHelper
         $index = self::getIncludesIndexColumnForSelect();
 
         return self::getBasePagesQuery($siteId)
-            ->select(['caches.id', 'uri', $index . ' AS index', 'template', 'params', 'elementCount', 'elementQueryCount', 'expiryDate'])
+            ->select(['caches.id', 'caches.siteId', 'uri', $index . ' AS index', 'template', 'params', 'elementCount', 'elementQueryCount', 'dateCached', 'expiryDate'])
             ->innerJoin([
                 'indexes' => IncludeRecord::find()
                     ->where(['siteId' => $siteId]),
@@ -642,7 +636,7 @@ class DiagnosticsHelper
     {
         $query = CacheRecord::find()
             ->from(['caches' => CacheRecord::tableName()])
-            ->select(['id', 'uri', 'elementCount', 'elementQueryCount', 'tagCount', 'expiryDate'])
+            ->select(['id', 'siteId', 'uri', 'elementCount', 'elementQueryCount', 'tagCount', 'dateCached', 'expiryDate'])
             ->leftJoin([
                 'elements' => ElementCacheRecord::find()
                     ->select(['cacheId', 'count(*) as elementCount'])