+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release Notes for Blitz
 
+## 4.22.0 - Unreleased
+
+> [!NOTE]
+> For the cache and expiry dates to appear in the new sidebar panel, the cache should be cleared or refreshed after this update completes.
+
+### Added
+
+- Added a sidebar panel to element edit pages ([#690](https://github.com/putyourlightson/craft-blitz/issues/690)).
+- Added a `dateCached` column to cache records which is output in the sidebar panel and the Blitz Diagnostics utility.
+
+### Changed
+
+- Blitz now requires Craft CMS 4.5.0 or later.
+
 ## 4.21.0 - 2024-07-24
 
 ### Added